feat: add favorite basic handler

+++ server/idl/api.thrift
@@ -98,7 +98,7 @@ struct douyin_publish_list_response {
 struct douyin_favorite_action_request {
     1: string token(api.query="token") // User authentication token
     2: i64 video_id(api.query="video_id", api.vd="len($)>0 && len($)<20") // Video Id
-    3: i32 action_type(api.query="action_type", api.vd="$==1 || $==2") // 1-like, 2-unlike
+    3: i8 action_type(api.query="action_type", api.vd="$==1 || $==2") // 1-like, 2-unlike
 }
 
 struct douyin_favorite_action_response {
@@ -120,7 +120,7 @@ struct douyin_favorite_list_response {
 struct douyin_comment_action_request {
     1: string token(api.query="token") // User authentication token
     2: i64 video_id(api.query="video_id", api.vd="len($)>0 && len($)<20") // Video Id
-    3: i32 action_type(api.query="action_type", api.vd="$==1 || $==2") // 1-like, 2-unlike
+    3: i8 action_type(api.query="action_type", api.vd="$==1 || $==2") // 1-like, 2-unlike
     4: string comment_text(api.query="comment_text") // The content of the comment filled by the user, used when action_type=1
     5: i64 comment_id(api.query="comment_id", api.vd="len($)>0 && len($)<20") // The comment id to be deleted is used when action_type=2
 }
@@ -145,7 +145,7 @@ struct douyin_comment_list_response {
 struct douyin_relation_action_request {
     1: string token(api.query="token") // User authentication token
     2: i64 to_user_id(api.query="to_user_id", api.vd="len($)>0 && len($)<20") // The other party's user id
-    3: i32 action_type(api.query="action_type", api.vd="$==1 || $==2") // 1-Follow, 2-Unfollow
+    3: i8 action_type(api.query="action_type", api.vd="$==1 || $==2") // 1-Follow, 2-Unfollow
 }
 
 struct douyin_relation_action_response {
