+++ server/idl/sociality.thrift
@@ -16,7 +16,7 @@ struct douyin_base_response{
 struct douyin_relation_action_request {
     1: i64 user_id // User Id
     2: i64 to_user_id // The other party's user id
-    3: i32 action_type // 1-Follow, 2-Unfollow
+    3: i8 action_type // 1-Follow, 2-Unfollow
 }
 
 struct douyin_relation_action_response {
